Videos/2026-03-10_Hypothetical_Claude_Code_Loop_Build_247_AI_Agents_2026_Conce/requirements.txt
anthropic>=0.40.0
PyYAML>=6.0